startup_project/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Django==5.2.5
fastapi==0.116.1
h11==0.16.0
idna==3.10
numpy==2.3.2
pydantic==2.11.7
pydantic_core==2.33.2
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
requests==2.32.4
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
python-dotenv==1.1.1
